refactor: extract portfolio value calculations into utility functions for consistency and clarity
